make nature_classic aware of moretrees
If moretrees is not present, nature_classic spawns blossoms and
eventually apples on default trees, as it always has.

If moretrees IS present, the code will only spawn apple blossoms on
moretrees' own apple tree leaves, and will only grow new apples on those
trees, not on default trees.  In this case, the nature:blossom node is
considered deprecated and is aliased to default:leaves to avoid causing
undefined nodes in old maps.