Regular & giant treefern leaves: always drop a sapling, and either leaves or another sapling
As there is only one leaf per fern, this makes it possible and convenient
to farm ferns (which was impossible, and even replanting all chopped-down
fern plants was extremely laborious due to the rarity of saplings).

Although on average slightly more than one sapling is produced per plant,
replanting and reharvesting the leaves is still the quickest way to
obtain more saplings.